requirements.txt
flask
flask-restful
pykwalify
GitPython
pyyaml

